Drop the TabletOwnership trait and replace with into_iterator()
Instead of messing with the ownership, make the Cache the main
entrypoint and provide iter() and into_iter() for the cache. The styli
are never into_iter()'ed anyway (except debugging tools which we don't
need to optimize for).

This also drops the find_tablet/take_tablet code and punts that to the
callers to do via filter() and into_iterator(). This isn't C here so
this API wasn't really needed.