Create a new Tool enum, move erasers there
Instead of mapping the (organically grown) data files as-is let's
separate the tools via a Tool enum, containing Stylus and
StylusWithEraser - the latter gives us a natural pairing
for the stylus + eraser and where the caller cares about it, they get
both.

This removes the need for the PairedId in the API that requires the
caller to look up the eraser, and the libwacom mess with "is this an
eraser? does it have an eraser?" which is not straightforward.